Move AdditionalJpaModelBuildItem to an SPI module
So that it may be used by smallrye-reactive-messaging, for whom the
Hibernate ORM extension is just an optional dependency.
