+++ extensions/hibernate-orm/deployment-spi/src/main/java/io/quarkus/hibernate/orm/deployment/spi/AdditionalJpaModelBuildItem.java
@@ -1,0 +1,24 @@
+package io.quarkus.hibernate.orm.deployment.spi;
+
+import java.util.Objects;
+
+import io.quarkus.builder.item.MultiBuildItem;
+
+/**
+ * Additional Jpa model class that we need to index
+ *
+ * @author Stéphane Épardaud
+ */
+public final class AdditionalJpaModelBuildItem extends MultiBuildItem {
+
+    private final String className;
+
+    public AdditionalJpaModelBuildItem(String className) {
+        Objects.requireNonNull(className);
+        this.className = className;
+    }
+
+    public String getClassName() {
+        return className;
+    }
+}
